+++ lang/en/local_boostnavigation.php
@@ -25,9 +25,6 @@
 defined('MOODLE_INTERNAL') || die();
 
 $string['pluginname'] = 'Boost navigation fumbling';
-$string['setting_currentcoursepresentation'] = 'Improve the current course\'s presentation in Boost\'s nav drawer';
-$string['setting_currentcoursefullname'] = 'Switch shortname to fullname';
-$string['setting_currentcoursefullname_desc'] = 'Enabling this setting will show the course\'s fullname instead of the course\'s shortname in Boost\'s nav drawer.<br />Technically, this can currently only be done by replacing the shortname globally at runtime, thus enabling this setting might have side effects to other parts of Moodle.';
 $string['setting_removenodesheading'] = 'Remove root nodes from Boost\'s nav drawer';
 $string['setting_removenodestechnicalhint'] = 'Technically, this is done by setting the node\'s showinflatnavigation attribute to false. Thus, the node will only be hidden from the nav drawer, but it will remain in the navigation tree and can still be accessed by other parts of Moodle.';
 $string['setting_removecalendarnode'] = 'Remove "Calendar" node';